requirements.txt
python-telegram-bot==20.7
qrcode==7.4.2
Pillow==10.1.0
mercadopago==2.2.0
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.9.1
cryptography==41.0.7
python-dateutil==2.8.2
requests==2.31.0
flask==2.3.3 